Updated iec104.cpp & .h files to fix sonar errors

+++ include/iec104.h
@@ -72,7 +72,7 @@ private:
 
     static std::string m_checkExchangedDataLayer(unsigned int ca,
                                                  const std::string& type_id,
-                                                 unsigned int ioa);
+                                                 uint64_t ioa);
 
     static CS104_Connection m_createTlsConnection(const char* ip, int port);
